+++ lib/braintree/credit_card_gateway.rb
@@ -79,7 +79,7 @@ module Braintree
 
     def self._signature(type) # :nodoc:
       billing_address_params = AddressGateway._shared_signature
-      options = [:make_default, :verification_merchant_account_id, :verify_card, :verification_amount, :venmo_sdk_session, :fail_on_duplicate_payment_method, :verification_account_type]
+      options = [:make_default, :verification_merchant_account_id, :verify_card, :verification_amount, :venmo_sdk_session, :fail_on_duplicate_payment_method, :verification_account_type,:verification_currency_iso_code]
       signature = [
         :billing_address_id, :cardholder_name, :cvv, :expiration_date, :expiration_month,
         :expiration_year, :number, :token, :venmo_sdk_payment_method_code, :device_data,