+++ spec/integration/braintree/payment_method_us_bank_account_spec.rb
@@ -13,7 +13,8 @@ describe Braintree::PaymentMethod do
       context "plaid verified nonce" do
         let(:nonce) { generate_valid_plaid_us_bank_account_nonce }
 
-        it "succeeds" do
+        # we are temporarily skipping this test until we have a more stable CI env
+        xit "succeeds" do
           customer = Braintree::Customer.create.customer
           result = Braintree::PaymentMethod.create(
             :payment_method_nonce => nonce,
@@ -49,7 +50,8 @@ describe Braintree::PaymentMethod do
           Braintree::UsBankAccountVerification::VerificationMethod::IndependentCheck,
           Braintree::UsBankAccountVerification::VerificationMethod::NetworkCheck,
         ].each do |method|
-          it "succeeds and verifies via #{method}" do
+          # we are temporarily skipping this test until we have a more stable CI env
+          xit "succeeds and verifies via #{method}" do
             customer = Braintree::Customer.create.customer
             result = Braintree::PaymentMethod.create(
               :payment_method_nonce => nonce,
@@ -183,7 +185,8 @@ describe Braintree::PaymentMethod do
 
         let(:nonce) { generate_valid_plaid_us_bank_account_nonce }
 
-        it "succeeds and verifies via independent check" do
+        # we are temporarily skipping this test until we have a more stable CI env
+        xit "succeeds and verifies via independent check" do
           customer = Braintree::Customer.create.customer
           result = Braintree::PaymentMethod.create(
             :payment_method_nonce => nonce,
@@ -219,7 +222,8 @@ describe Braintree::PaymentMethod do
       context "non plaid verified nonce" do
         let(:nonce) { generate_non_plaid_us_bank_account_nonce }
 
-        it "succeeds and verifies via independent check" do
+        # we are temporarily skipping this test until we have a more stable CI env
+        xit "succeeds and verifies via independent check" do
           customer = Braintree::Customer.create.customer
           result = Braintree::PaymentMethod.create(
             :payment_method_nonce => nonce,