+++ spec/spec_helper.rb
@@ -39,6 +39,7 @@ unless defined?(SPEC_HELPER_LOADED)
     AnotherUsBankMerchantAccountId = "another_us_bank_merchant_account"
     AdyenMerchantAccountId = "adyen_ma"
     HiperBRLMerchantAccountId = "hiper_brl"
+    CardProcessorBRLMerchantAccountId = "card_processor_brl"
 
     TrialPlan = {
       :description => "Plan for integration tests -- with trial",